User Auth

+++ app.py
@@ -1,27 +1,62 @@
-from crypt import methods
+# Imports
+from distutils.log import error
 from flask import Flask, render_template, request, redirect
 from flask_sqlalchemy import SQLAlchemy
+from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import UserMixin, LoginManager, login_required, current_user, login_user, logout_user
 from datetime import datetime
 
+
+# Configs
 app = Flask(__name__)
+app.secret_key = 'xyz'
+
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
 db = SQLAlchemy(app)
 
+login = LoginManager()
+login.init_app(app)
+login.login_view = 'login'
 
+# Models
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     content = db.Column(db.String(200), nullable=False)
     date_created = db.Column(db.DateTime, default=datetime.utcnow)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    user = db.relationship('UserModel')
 
     def __repr__(self):
         return '<Task %r>' % self.id
 
 
+class UserModel(UserMixin, db.Model):
+    __tablename__ = 'users'
+ 
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(80), unique=True)
+    username = db.Column(db.String(100))
+    password_hash = db.Column(db.String())
+    todo = db.relationship('Todo', cascade="all,delete", back_populates='user', lazy='dynamic')
+ 
+    def set_password(self,password):
+        self.password_hash = generate_password_hash(password)
+     
+    def check_password(self,password):
+        return check_password_hash(self.password_hash,password)
+
+@login.user_loader
+def load_user(id):
+    return UserModel.query.get(int(id))
+
+
+# Routes
 @app.route('/', methods=['GET', 'POST'])
+@login_required
 def index():
     if request.method == 'POST':
         task_content = request.form['content']
-        new_task = Todo(content=task_content)
+        new_task = Todo(content=task_content, user=current_user)
         try:
             db.session.add(new_task)
             db.session.commit()
@@ -29,38 +64,93 @@ def index():
         except:
             return 'There was an issue adding your task'
     else:
-        tasks = Todo.query.order_by(Todo.date_created).all()
+        tasks = Todo.query.filter_by(user = current_user).order_by(Todo.date_created).all()
         return render_template('index.html', tasks=tasks)
 
 
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
     task_to_delete = Todo.query.get_or_404(id)
 
-    try:
-        db.session.delete(task_to_delete)
-        db.session.commit()
-        return redirect('/')
-    except:
-        return 'There was a problem deleting that task'
+    if task_to_delete.user == current_user:
+        try:
+            db.session.delete(task_to_delete)
+            db.session.commit()
+            return redirect('/')
+        except:
+            return 'There was a problem deleting that task'
+
+    else:
+        return "Access Denied!"
 
 
 @app.route('/update/<int:id>', methods=['GET', 'POST'])
+@login_required
 def update(id):
     task = Todo.query.get_or_404(id)
 
     if request.method == 'POST':
         task.content = request.form['content']
-
         try:
             db.session.commit()
             return redirect('/')
         except:
             return 'There was an issue updating your task'
-
-    else:
+    
+    elif task.user == current_user:
         return render_template('update.html', task=task)
 
+    else:
+         return "Access Denied!"
+        
 
+@app.route('/login', methods = ['POST', 'GET'])
+def login():
+    error = ""
+    if current_user.is_authenticated:
+        return redirect('/')
+     
+    if request.method == 'POST':
+        email = request.form['email']
+        user = UserModel.query.filter_by(email = email).first()
+        if user is not None and user.check_password(request.form['password']):
+            login_user(user)
+            return redirect('/')
+        else:
+            error = "Invalid login detils"
+    
+    return render_template('login.html', error=error)
+
+
+@app.route('/register', methods=['POST', 'GET'])
+def register():
+    error = ""
+    if current_user.is_authenticated:
+        return redirect('/')
+     
+    if request.method == 'POST':
+        email = request.form['email']
+        username = request.form['username']
+        password = request.form['password']
+ 
+        if UserModel.query.filter_by(email=email).first():
+            error = 'Email already exists'
+            return render_template('register.html', error=error)
+             
+        user = UserModel(email=email, username=username)
+        user.set_password(password)
+        db.session.add(user)
+        db.session.commit()
+        return redirect('/login')
+    return render_template('register.html', error=error)
+
+
+@app.route('/logout')
+def logout():
+    logout_user()
+    return redirect('/')
+
+# Main
 if __name__ == "__main__":
     app.run(debug=True)
